Move image_Dataset to new file
